Fix issue with loading paired devices for device selection.
A bug was found in the code that handled the device select response received from the Garmin's Connect app that provided a list of paired devices in the Connect app. This bug was introduced when iOS behavior was changed for handling the loadUrl call to not have the source app parameter set when the calling app is not developed by the same team. This change was intoduced since iOS13. Fixed the issue by not explicitly checking for the soruce app since the CIQ SDK developed by third party developers will always have the source app parameter set to nil. The specific option in question that changed the behavior is UIApplicationOpenURLOptionsSourceApplicationKey(https://developer.apple.com/documentation/uikit/uiapplicationopenurloptionssourceapplicationkey?language=objc)